in progress

+++ src/com/training/Spring_BankingApp/src/main/java/com/day7/repository/CustomerDao.java
@@ -8,6 +8,6 @@ public interface CustomerDao {
 	
 	public int viewCustomer(String customer_id);
 	
-	public int performTransaction(String user_id,String receiver_id,double amount); 
+	public int performTransaction(String user_id,String receiver_id,int amount); 
 	
 }
